[QOL-8544] rename module to avoid clash
